Phase 11.5
Added a new class to calculate the number of existing customer accounts. And implemented this to the 'NEW Customer' panel in the 'Teller' frame. Now before registering a new account it will check if the customer has exceeded above the maximum number of customer accounts (five). If so, a message box would be be shown or otherwise the process would continue.

Co-Authored-By: Ranul <lranul@users.noreply.github.com>
Co-Authored-By: nnwijesuriya <nnwijesuriya@users.noreply.github.com>
Co-Authored-By: IshanC7 <thenooobc7@users.noreply.github.com>
Co-Authored-By: ryancargan <ryancargan@users.noreply.github.com>